Make the styleguide more generic
